Plenty new image resize functionality & Bugfixes
The images now doesn’t get strecht or anything to fit in the
imageslider. The new idea of it is to look after the ration of the
image, and crop it instead to focus on the middle and not to produce
such gray spaces like other products. If you doesn’t like it, just
enjoy the next version wheres also a ugly greyspace version in it!