Add new navigation type "button"
This navigation type kills the normal navigation (in each skin) and
stops the automatic animation. Instead of it you get two buttons
left/right to navigate through all pictures in a loop. The rest of the
functions like link referrer still works. If you put in a lightbox you
can use it as an image viewer.